pyproject.toml
[tool.poetry]
name = "warehouse_monitoring"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
pre-commit = "^2.20.0"
djangorestframework = "^3.14.0"
pandas = "^1.5.0"
openpyxl = "^3.0.10"
drf-yasg = "^1.21.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
